fix: move RunContext out of worker types
